callpilot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
jinja2==3.1.5
python-multipart==0.0.20
retell-sdk==4.16.0
pydantic==2.10.4
pydantic-settings==2.7.1
aiofiles==24.1.0
stripe>=8.0.0
openai>=1.0.0
